requirements.txt
cloudpickle==0.6.1
cycler==0.10.0
dask==1.0.0
decorator==4.3.0
kiwisolver==1.0.1
matplotlib==3.0.2
networkx==2.2
numpy==1.15.4
Pillow==5.3.0
pyparsing==2.3.0
python-dateutil==2.7.5
PyWavelets==1.0.1
scikit-image==0.14.1
scipy==1.1.0
six==1.11.0
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
Django==2.2.2
gunicorn==19.6.0




